+++ kotlin-stdlib/common/test/kotlin/com/javiersc/kotlin/stdlib/CharSequenceTest.kt
@@ -69,30 +69,4 @@ internal class CharSequenceTest {
         assertFalse { "Hello World 1".notContain(regex) }
         assertTrue { "Hello World".notContain(regex) }
     }
-
-    @Test
-    fun string_remove_duplicate_empty_lines() {
-        assertTrue { "a\nb\n\n\nc\n".removeDuplicateEmptyLines() == "a\nb\n\nc\n" }
-        assertTrue { "a\n\nb\n\n\nc\n".removeDuplicateEmptyLines() == "a\n\nb\n\nc\n" }
-        assertTrue { "a\n\nb\n\n\nc\n\n".removeDuplicateEmptyLines() == "a\n\nb\n\nc\n" }
-        assertTrue { "a\n\n\n\n\nb\n\n".removeDuplicateEmptyLines() == "a\n\nb\n" }
-    }
-
-    @Test
-    fun string_end_with_new_line() {
-        assertTrue { "a".endWithNewLine() == "a\n" }
-        assertTrue { "a\n".endWithNewLine() == "a\n" }
-        assertTrue { "".endWithNewLine() == "" }
-        assertTrue { "a\nb".endWithNewLine() == "a\nb\n" }
-        assertTrue { "".endWithNewLine() == "" }
-        assertTrue { "\n".endWithNewLine() == "\n" }
-
-        val helloWorld =
-            """
-                |Hello World
-                |
-            """
-
-        assertTrue { helloWorld.trimMargin().endWithNewLine() == helloWorld.trimMargin() }
-    }
 }